Create `utils::registered_metric` module and move code over
