+++ node/src/utils.rs
@@ -8,6 +8,7 @@ mod external;
 pub(crate) mod fmt_limit;
 mod fuse;
 pub(crate) mod opt_display;
+pub(crate) mod registered_metric;
 pub(crate) mod rlimit;
 pub(crate) mod round_robin;
 pub(crate) mod umask;
@@ -31,11 +32,7 @@ use fs2::FileExt;
 use futures::future::Either;
 use hyper::server::{conn::AddrIncoming, Builder, Server};
 
-use prometheus::{
-    self,
-    core::{Atomic, Collector, GenericCounter},
-    Histogram, HistogramOpts, IntCounter, IntGauge, Registry,
-};
+use prometheus::{self, Histogram, HistogramOpts, IntGauge, Registry};
 use serde::Serialize;
 use thiserror::Error;
 use tracing::{error, warn};
@@ -491,98 +488,6 @@ impl<A, B, F, G> Peel for Either<(A, G), (B, F)> {
             Either::Left((v, _)) => Either::Left(v),
             Either::Right((v, _)) => Either::Right(v),
         }
-    }
-}
-
-#[derive(Debug)]
-pub(crate) struct RegisteredMetric<T>
-where
-    T: Collector + 'static,
-{
-    metric: Option<Box<T>>,
-    registry: Registry,
-}
-
-impl<T> RegisteredMetric<T>
-where
-    T: Collector + 'static,
-{
-    pub(crate) fn new(registry: Registry, metric: T) -> Result<Self, prometheus::Error>
-    where
-        T: Clone,
-    {
-        let boxed_metric = Box::new(metric);
-        registry.register(boxed_metric.clone())?;
-
-        Ok(RegisteredMetric {
-            metric: Some(boxed_metric),
-            registry,
-        })
-    }
-
-    #[inline]
-    pub(crate) fn inner(&self) -> &T {
-        self.metric.as_ref().expect("metric disappeared")
-    }
-}
-
-impl<P> RegisteredMetric<GenericCounter<P>>
-where
-    P: Atomic,
-{
-    #[inline]
-    pub(crate) fn inc(&self) {
-        self.inner().inc()
-    }
-}
-
-impl<T> Drop for RegisteredMetric<T>
-where
-    T: Collector + 'static,
-{
-    fn drop(&mut self) {
-        if let Some(boxed_metric) = self.metric.take() {
-            let desc = boxed_metric
-                .desc()
-                .iter()
-                .next()
-                .map(|desc| desc.fq_name.clone())
-                .unwrap_or_default();
-            self.registry.unregister(boxed_metric).unwrap_or_else(|_| {
-                tracing::error!("unregistering {} failed: was not registered", desc)
-            })
-        }
-    }
-}
-
-pub(crate) trait RegistryExt {
-    fn new_int_counter<S1: Into<String>, S2: Into<String>>(
-        &self,
-        name: S1,
-        help: S2,
-    ) -> Result<RegisteredMetric<IntCounter>, prometheus::Error>;
-    fn new_int_gauge<S1: Into<String>, S2: Into<String>>(
-        &self,
-        name: S1,
-        help: S2,
-    ) -> Result<RegisteredMetric<IntGauge>, prometheus::Error>;
-}
-
-impl RegistryExt for Registry {
-    fn new_int_counter<S1: Into<String>, S2: Into<String>>(
-        &self,
-        name: S1,
-        help: S2,
-    ) -> Result<RegisteredMetric<IntCounter>, prometheus::Error> {
-        RegisteredMetric::new(self.clone(), IntCounter::new(name, help)?)
-    }
-
-    fn new_int_gauge<S1: Into<String>, S2: Into<String>>(
-        &self,
-        name: S1,
-        help: S2,
-    ) -> Result<RegisteredMetric<IntGauge>, prometheus::Error> {
-        RegisteredMetric::new(self.clone(), IntGauge::new(name, help)?)
     }
 }
